test: 패키지 변경 및 기능 추가에 따른 테스트 코드 작성
